+++ shared/lib_battery_dispatch.h
@@ -109,7 +109,7 @@ public:
 		double P_system_clipped = 0
 		) = 0;
 
-	virtual bool check_constraints(double &I, int count);
+	virtual bool check_constraints(double &I, size_t count);
 
 	battery_t * battery_model(){ return _Battery; }
 
@@ -267,7 +267,7 @@ protected:
 		std::map<size_t, double> dm_percent_gridcharge);
 
 	void SOC_controller();
-	bool check_constraints(double &I, int count);
+	bool check_constraints(double &I, size_t count);
 
 	util::matrix_t < size_t > _sched;
 	util::matrix_t < size_t > _sched_weekend;
@@ -370,7 +370,7 @@ public:
 	virtual void set_custom_dispatch(std::vector<double> P_batt_dc);
 
 	/* Check constraints and re-dispatch if needed */
-	virtual bool check_constraints(double &I, int count);
+	virtual bool check_constraints(double &I, size_t count);
 
 protected:
